% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gREM.R
\name{calcProfileWidth}
\alias{calcProfileWidth}
\title{Internal function to calculate profile width as described in the text}
\usage{
calcProfileWidth(alpha, theta, r)
}
\arguments{
\item{alpha}{Call width in radians.}

\item{theta}{Detector width in radians.}

\item{r}{Sensor detection radius in metres.}
}
\description{
Internal function to calculate profile width as described in the text
}
