% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gREM.R
\name{gremAbundance}
\alias{gremAbundance}
\title{Calculate abudance using the gREM.}
\usage{
gremAbundance(z, alpha, theta, r, animalSpeed, t, area)
}
\arguments{
\item{z}{Number of detections.}

\item{alpha}{Call width in radians.}

\item{theta}{Detector width in radians.}

\item{r}{Sensor detection radius in metres.}

\item{animalSpeed}{Average animal speed in metres per second.}

\item{t}{Total survey time. This is the amount of time the sensors are
active multiplied by the number of sensors used.}

\item{area}{The size of the study area in metres squared.}
}
\description{
Calculate abundance from count data using the gREM.
}
\details{
It is necessary to define a study area to calculate abundance rather than
   density. The easiest way to do this is to define the study area before
   data collection and then place camera traps/acoustic detectors/etc.
   randomly inside the area. It is more difficult to define the area
   studied post hoc.

Note that the REM of Rowcliffe et al. 2008 can be used by setting alpha to 
   zero and the Gas Model of Yapp 1956 can be used setting alpha and theta
   to zero.
}
\references{
Lucas, T. C. D., Moorcroft, E. A., Freeman, R., Rowcliffe, J. M.,
   Jones, K. E. (2015), A generalised random encounter model for estimating 
   animal density with remote sensor data. Methods in Ecology and Evolution. 
   doi: 10.1111/2041-210X.12346

Rowcliffe, J., Field, J., Turvey, S. & Carbone, C. (2008) Estimating animal 
   density using camera traps without the need for individual recognition. 
   Journal of Applied Ecology, 45, 1228-1236.

Yapp, W. (1956) The theory of line transects. Bird Study, 3, 93-104.
}
\seealso{
\code{\link{gremDensity}}
}
