% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gREM.R
\name{gremDensity}
\alias{gremDensity}
\title{Calculate population density using the gREM.}
\usage{
gremDensity(count, alpha, theta, r, v, tm)
}
\arguments{
\item{count}{Number of detections.}

\item{alpha}{Call width in radians.}

\item{theta}{Detector width in radians.}

\item{r}{Sensor detection radius in metres.}

\item{v}{Average animal speed in kilometers per day.}

\item{tm}{Total survey time in hours. This is the amount of time the sensors are
active multiplied by the number of sensors used.}
}
\description{
Calculate population density from count data using the gREM.
}
\details{
Note that the REM of Rowcliffe et al. 2008 can be used by setting alpha to 
   2*pi and the Gas Model of Yapp 1956 can be used setting alpha and theta
   to 2*pi.
}
\references{
Lucas, T. C. D., Moorcroft, E. A., Freeman, R., Rowcliffe, J. M.,
   Jones, K. E. (2015), A generalised random encounter model for estimating 
   animal density with remote sensor data. Methods in Ecology and Evolution. 
   doi: 10.1111/2041-210X.12346

Rowcliffe, J., Field, J., Turvey, S. & Carbone, C. (2008) Estimating animal 
   density using camera traps without the need for individual recognition. 
   Journal of Applied Ecology, 45, 1228-1236.

Yapp, W. (1956) The theory of line transects. Bird Study, 3, 93-104.
}
\seealso{
\code{\link{gremAbundance}}
}
