% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_data.R
\docType{data}
\name{hDat}
\alias{hDat}
\title{Example data for REM density estimation}
\format{An object of class \code{data.frame} with 118 rows and 4 columns.}
\source{
From data published as Caravaggi et al. (2016) Remote Sensing in Ecology and Conservation 2:45-58
(\href{http://onlinelibrary.wiley.com/doi/10.1002/rse2.11/abstract}{RSEC})
}
\usage{
data(hDat)
}
\description{
Simulated camera trap detection data for density estimation via Random Encounter Models (Rowcliffe et al, 2008). Real data used with the remBoot package should adhere to the same file structure in the first five columns - survey site, camera number, number of individuals in each capture, distance to capture, angle of detection.
}
\references{
Rowcliffe et al. (2008) Journal of Applied Ecology 45:1228-1236
(\href{http://onlinelibrary.wiley.com/doi/10.1111/j.1365-2664.2008.01473.x/full}{J Appl Ecol})
}
\keyword{datasets}

