% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dat.R
\name{split_dat}
\alias{split_dat}
\title{Data splitting for grpDat}
\usage{
split_dat(.df)
}
\arguments{
\item{.df}{= the data frame}
}
\description{
Separate data into groups prior to bootstrapping. Calls grpDat_func.
}
\examples{

## Read data:
data(hDat)

## Split data by survey site:
grpDat <- split_dat(hDat)

## Check that function has performed correctly:
lapply(grpDat, head)
lapply(grpDat, nrow)
}
