% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sd.R
\name{boot_sd}
\alias{boot_sd}
\title{Required for remsD bootstrapping}
\usage{
boot_sd(x, tm, v, nboots)
}
\arguments{
\item{x}{bsD resampled data}

\item{tm}{(numeric) The total number of hours all cameras were left in-situ at a focal site}

\item{v}{(numeric) The distance travelled by the focal species in 24 hours, in kilometres}

\item{nboots}{(numeric) The number of bootstrap iterations}
}
\description{
Resamples camera trap data to calculate variance for Random Encounter Model density estimates. Note that tm and v must be the same across all sites. If this is not the case, split your data before bootstrapping and run the function on each sub-group. Called by remsD. Calls bsD.
}
\examples{
## Define the number of bootstrapping iterations and apply boot_sd to the data:
tm <- 3600
v <- 0.85
nboots <- 1000

grpDat <- split_dat(hDat)
remsD <- lapply(grpDat, boot_sd, tm, v)
remsSD <- lapply(remsD, sd)
remsSD

}
